Include the manic/reverse flags in Supreme sound.cpp
These don't work yet because the new sound engine doesn't support them,
but this is part of making sure they will work.

+++ source/ham/jamulsound.cpp
@@ -102,7 +102,7 @@ void JamulSoundUpdate(void)
 	}
 }
 
-bool JamulSoundPlay(int which,long pan,long vol,byte playFlags,int priority)
+bool JamulSoundPlay(int which,long pan,long vol,int playFlags,int priority)
 {
 	char s[32];
 	int i,chosen,lowpriority;
@@ -253,7 +253,7 @@ void JamulSoundPurge(void)
 	}
 }
 
-void GoPlaySound(int num,long pan,long vol,byte flags,int priority)
+void GoPlaySound(int num,long pan,long vol,int flags,int priority)
 {
 	if(!soundIsOn)
 		return;
